Use a more portable date construct.
The command:

    date --rfc-3339=seconds

is a GNU extension.  A much more portable date comamnd:

    date +"%Y-%m-%d %H:%M:%S%z"

uses common strftime(3) format specifiers, works on Linux, BSD, and macOS,
and produces output nearly identical to the first form (differs only in
the lack of a colon between the hours and minutes of the timezone offset
field).